ViewHooks: add a compile time option (default off) to have dynamic view hooks as the default. Refactor some code to prevent noncopyable views from generating deep_copy calls in ViewHolder
