+++ core/src/Kokkos_View.hpp
@@ -524,6 +524,7 @@ constexpr bool is_assignable(const Kokkos::View<ViewTDst...>& dst,
 
 #include <impl/Kokkos_ViewMapping.hpp>
 #include <impl/Kokkos_ViewArray.hpp>
+#include <View/Hooks/Kokkos_DynamicViewHooks.hpp>
 
 //----------------------------------------------------------------------------
 //----------------------------------------------------------------------------
