+++ core/src/Kokkos_ViewHolder.hpp
@@ -45,6 +45,7 @@
 #ifndef KOKKOS_VIEWHOLDER_HPP
 #define KOKKOS_VIEWHOLDER_HPP
 
+#include <Kokkos_Core_fwd.hpp>
 #include <Kokkos_Macros.hpp>
 #include <Kokkos_HostSpace.hpp>
 #include <Kokkos_Layout.hpp>
@@ -54,6 +55,9 @@
 #include <type_traits>
 
 namespace Kokkos {
+template <class T1, class T2>
+struct is_always_assignable_impl;
+
 template <class DataType, class... Properties>
 class View;
 
@@ -64,12 +68,12 @@ namespace Experimental {
 namespace Impl {
 
 template <class DataType, class... Properties>
-View<typename ViewTraits<DataType, Properties...>::non_const_data_type,
-     typename std::conditional<
-         std::is_same<LayoutLeft, typename ViewTraits<DataType, Properties...>::
-                                      array_layout>::value,
-         LayoutLeft, LayoutRight>::type,
-     HostSpace, MemoryTraits<Unmanaged>>
+using unmanaged_view_type_like = View<typename ViewTraits<DataType, Properties...>::non_const_data_type,
+    typename ViewTraits<DataType, Properties...>::array_layout,
+    HostSpace, MemoryTraits<Unmanaged > >;
+
+template <class DataType, class... Properties>
+unmanaged_view_type_like< DataType, Properties... >
 make_unmanaged_view_like(View<DataType, Properties...> view,
                          unsigned char *buff) {
   static_assert(
@@ -86,8 +90,7 @@ make_unmanaged_view_like(View<DataType, Properties...> view,
       std::is_same<LayoutLeft, typename traits_type::array_layout>::value,
       LayoutLeft, LayoutRight>::type;
 
-  using new_view_type =
-      View<new_data_type, layout_type, HostSpace, MemoryTraits<Unmanaged>>;
+  using new_view_type = unmanaged_view_type_like< DataType, Properties... >;
 
   return new_view_type(
       reinterpret_cast<typename new_view_type::pointer_type>(buff),
@@ -172,6 +175,42 @@ class ViewHolderImplBase {
   bool m_is_hostspace     = false;
 };
 
+template< typename SrcViewType, typename DstViewType, typename Enabled = void >
+struct ViewHolderImplDeepCopyImpl
+{
+  static void copy_to_unmanaged( SrcViewType &, void * )
+  {
+    Kokkos::Impl::throw_runtime_exception(
+        "Cannot deep copy a view holder to an incompatible view" );
+  }
+
+  static void copy_from_unmanaged( DstViewType &, const void * )
+  {
+    Kokkos::Impl::throw_runtime_exception(
+        "Cannot deep copy from a host unmanaged view holder to an incompatible view" );
+  }
+};
+
+template< typename SrcViewType, typename DstViewType >
+struct ViewHolderImplDeepCopyImpl< SrcViewType, DstViewType,
+    std::enable_if_t< is_always_assignable_impl<
+      typename std::remove_reference<DstViewType>::type,
+    typename std::remove_const<
+    typename std::remove_reference<SrcViewType>::type>::type>::value > >
+{
+  static void copy_to_unmanaged( SrcViewType &_src, void *_buff )
+  {
+    auto dst = make_unmanaged_view_like(_src, _buff);
+    deep_copy(dst, _src);
+  }
+
+  static void copy_from_unmanaged( DstViewType &_dst, const void *_buff )
+  {
+    auto src = Impl::make_unmanaged_view_like(_dst, _buff);
+    deep_copy(_dst, src);
+  }
+};
+
 template <typename View, typename Enable = void>
 class ViewHolderImpl : public ViewHolderImplBase {
   static_assert(
@@ -189,13 +228,13 @@ class ViewHolderImpl : public ViewHolderImplBase {
         m_view(view) {}
 
   void deep_copy_to_buffer(unsigned char *buff) override {
-    auto unmanaged = Impl::make_unmanaged_view_like(m_view, buff);
-    deep_copy(unmanaged, m_view);
+    using dst_type = unmanaged_view_type_like< View >;
+    ViewHolderImplDeepCopyImpl< View, dst_type >::copy_to_unmanaged(m_view, buff);
   }
 
   void deep_copy_from_buffer(unsigned char *buff) override {
-    auto unmanaged = Impl::make_unmanaged_view_like(m_view, buff);
-    deep_copy(m_view, unmanaged);
+    using src_type = unmanaged_view_type_like< View >;
+    ViewHolderImplDeepCopyImpl< src_type, View >::copy_from_unmanaged(m_view, buff);
   }
 
   ViewHolderImpl *clone() const override { return new ViewHolderImpl(m_view); }
@@ -223,8 +262,8 @@ class ViewHolderImpl<View, typename std::enable_if<std::is_const<
         m_view(view) {}
 
   void deep_copy_to_buffer(unsigned char *buff) override {
-    auto unmanaged = Impl::make_unmanaged_view_like(m_view, buff);
-    deep_copy(unmanaged, m_view);
+    using dst_type = unmanaged_view_type_like< View >;
+    ViewHolderImplDeepCopyImpl< View, dst_type >::copy_to_unmanaged(m_view, buff);
   }
 
   ViewHolderImpl *clone() const override { return new ViewHolderImpl(m_view); }
